task six

+++ src/Entities/Assessment.java
@@ -2,23 +2,23 @@ package Entities;
 
 public class Assessment{
 
-    private int idAssessment;
+    private Long idAssessment;
     private int grade; //nota
     private Order order;
     private Restaurant restaurant;
 
-    public Assessment(int idAssessment, int grade, Order order, Restaurant restaurant) {
+    public Assessment(Long idAssessment, int grade, Order order, Restaurant restaurant) {
         this.idAssessment = idAssessment;
         this.grade = grade;
         this.order = order;
         this.restaurant = restaurant;
     }
 
-    public int getIdAssessment() {
+    public Long getIdAssessment() {
         return idAssessment;
     }
 
-    public void setIdAssessment(int idAssessment) {
+    public void setIdAssessment(Long idAssessment) {
         this.idAssessment = idAssessment;
     }
